backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Git repository documentation and analysis backend"
requires-python = ">=3.12"
dependencies = [
    "beanie>=1.30.0",
    "codetext>=0.0.9",
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.95.0",
    "fastapi-jwt-auth>=0.5.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.24.0",
    "litellm>=1.74.0",
    "matplotlib>=3.7.1",
    "motor>=3.7.1",
    "networkx>=3.1",
    "pydantic>=1.10.7",
    "pyjwt>=1.7.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.6",
    "pyvis>=0.3.2",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-typescript>=0.23.2",
    "urllib3<2.0",
    "uvicorn>=0.21.1",
    # Observability
    "arize-phoenix-otel",
    "openinference-instrumentation-litellm",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "routes*", "schemas*", "storage*", "graphing*", "controllers*", "documentationo_generator*"]
